+++ src/main/java/kitchenpos/order/delivery/application/OrderTableService.java
@@ -1,9 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.order.delivery.application;
 
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
+import kitchenpos.order.delivery.domain.OrderRepository;
+import kitchenpos.order.delivery.domain.OrderStatus;
+import kitchenpos.order.delivery.domain.OrderTable;
+import kitchenpos.order.delivery.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
