+++ src/main/java/kitchenpos/order/delivery/domain/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.delivery.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
