+++ src/main/java/kitchenpos/order/delivery/domain/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.delivery.domain;
 
 import java.util.List;
 import java.util.Optional;
