+++ src/main/java/kitchenpos/order/delivery/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.delivery.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
