+++ src/main/java/kitchenpos/order/delivery/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.delivery.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
