+++ src/main/java/kitchenpos/order/delivery/infra/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.order.delivery.infra;
 
 import java.math.BigDecimal;
 import java.util.UUID;
