+++ src/test/java/kitchenpos/Fixtures.java
@@ -4,6 +4,9 @@ import kitchenpos.domain.*;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.order.delivery.domain.*;
+import kitchenpos.order.domain.*;
+import kitchenpos.order.eatin.domain.*;
 import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
