+++ src/test/java/kitchenpos/application/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
+import kitchenpos.order.delivery.domain.OrderTable;
+import kitchenpos.order.delivery.domain.OrderTableRepository;
 
 import java.util.*;
