+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -2,6 +2,10 @@ package kitchenpos.application;
 
 import kitchenpos.domain.*;
 import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.order.delivery.application.OrderService;
+import kitchenpos.order.delivery.domain.*;
+import kitchenpos.order.domain.*;
+import kitchenpos.order.eatin.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
